Login con cookie funcionando

+++ Blog/borrar_mensaje.php
@@ -1,8 +1,10 @@
 <?php 
+session_start();
 
 require_once 'modelos/ConnexionDB.php';
 require_once 'modelos/Mensaje.php';
 require_once 'modelos/MensajesDAO.php';
+require_once 'funciones.php';
 
 //Creamos la conexión utilizando la clase que hemos creado
 $connexionDB = new ConnexionDB('root','','localhost','blog');
@@ -13,6 +15,15 @@ $mensajesDAO = new MensajesDAO($conn);
 
 //Obtener el mensaje
 $idMensaje = htmlspecialchars($_GET['id']);
-$mensajesDAO->delete($idMensaje);
+$mensaje = $mensajesDAO->getById($idMensaje);
+
+//Comprobamos que mensaje pertenece al usuario conectado
+if($_SESSION['id']==$mensaje->getIdUsuario()){
+    $mensajesDAO->delete($idMensaje);
+}
+else
+{
+    guardarMensaje("No puedes borrar este mensaje");
+}
 
 header('location: index.php');
